Refactoring : CowDogPageBody
